Move passbuilder into core.
